refactor: update AgentBookmarkButton and replace BackToGallery with ActionButtons in AgentDetail
